Update dcgs.pl & clpz.pl

+++ src/internal.h
@@ -512,7 +512,7 @@ struct trail_ {
 
 struct slot_ {
 	cell c;
-	uint32_t vgen, vgen2;
+	uint16_t vgen, vgen2;
 };
 
 // Where 'prev_offset' is the number of frames back
@@ -671,7 +671,7 @@ struct query_ {
 	pl_idx q_size[MAX_QUEUES], tmpq_size[MAX_QUEUES], qp[MAX_QUEUES];
 	prolog_flags flags;
 	enum q_retry retry;
-	uint32_t vgen;
+	uint16_t vgen;
 	int8_t halt_code;
 	int8_t quoted;
 	enum { WAS_OTHER, WAS_SPACE, WAS_COMMA, WAS_SYMBOL } last_thing;
